% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mlmm.gwas.AD}
\alias{mlmm.gwas.AD}
\alias{floweringDateAD}
\alias{Xa}
\alias{Xd}
\alias{K.add}
\alias{K.dom}
\title{Dataset for examples with the mlmm.gwas package, additive and additive+dominance models}
\format{\itemize{
  \item floweringDateAD: a named numeric of length 444.
  \item Xa: a 110x500 numeric matrix
  \item Xd: a 110x500 numeric matrix
  \item K.add: a 110x110 numeric matrix
  \item K.dom: a 110x110 numeric matrix
}}
\source{
\href{https://link.springer.com/article/10.1007/s00122-017-3003-4}{Bonnafous & al. (2017)}
}
\usage{
data(mlmm.gwas.AD)
}
\description{
\itemize{
   \item Species: Helianthus annuus
   \item Individuals: 125
   \item Markers: 500
}
}
\details{
Variables:
\itemize{
  \item floweringDateAD: flowering dates in °C.day, time since sowing.
  \item Xa: genotype matrix (additive)
  \item Xd: genotype matrix (dominance)
  \item K.add: "kinship" matrix (additive)
  \item K.dom: "kinship" matrix (dominance)
}
}
\keyword{dataset}
